feat: Update AppBarToggleButton to winui3/release/1.6.4
