+++ src/Uno.UI/UI/Xaml/Controls/AppBar/AppBarToggleButton.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Microsoft.UI.Xaml.Controls;
+
+/// <summary>
+/// Represents a button control that can switch states and be displayed in an AppBar.
+/// </summary>
+public partial class AppBarToggleButton
+{
+}
