+++ src/Uno.UI/UI/Xaml/Controls/CommandBar/ICommandBarLabeledElement.cs
@@ -7,6 +7,9 @@ namespace Microsoft.UI.Xaml.Controls
 	internal interface ICommandBarLabeledElement
 	{
 		void SetDefaultLabelPosition(CommandBarDefaultLabelPosition defaultLabelPosition);
+
 		bool GetHasBottomLabel();
+
+		bool GetHasRightLabel();
 	}
 }
